+++ rapdev_zoom/datadog_checks/rapdev_zoom/__about__.py
@@ -1,3 +1,3 @@
 # (C) RapDev, Inc. 2020-present
 # All rights reserved
-__version__ = '5.3.0'
+__version__ = '6.0.0'