+++ rapdev_zoom/datadog_checks/rapdev_zoom/helpers.py
@@ -1,20 +1,3 @@
-import jwt
-import time
-
-
-def generate_token(api_key, api_secret):
-    token = jwt.encode(
-        {"iss": api_key, "exp": time.time() + 60},
-        api_secret,
-        algorithm='HS256'
-    )
-    
-    if float(jwt.__version__.split('.')[0]) < 2:
-        token = token.decode('utf-8')
-
-    return token
-
-
 VALID_METRICS = [
     "bitrate",
     "avg_loss",